Extract llm setting
